backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
pdfplumber
spacy